credit

+++ BankingSystem/BankingSystem/bank_account.cs
@@ -17,7 +17,7 @@ namespace BankingSystem
         public int id { get; set; }
         public string Number { get; set; }
         public string User_login { get; set; }
-        public Nullable<int> Sum { get; set; }
+        public Nullable<double> Sum { get; set; }
     
         public virtual bank_deposit bank_deposit { get; set; }
         public virtual user user { get; set; }
